pyproject.toml
[project]
name = "asanappeal-ai"
version = "0.1.0"
description = "A runnable AI backend foundation for ASANAppeal."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.29.0",
  "pydantic>=2.6.0",
  "google-genai>=1.0.0",
  "openai>=1.0.0",
  "pillow>=12.0.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
